feat: timeline styling
