+++ .gitignore
@@ -7,3 +7,4 @@ public/
 static/admin/*.bundle.*
 .DS_Store
 yarn-error.log
+/.idea
